Moved input-wrappers
