+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/autocomplete/autocomplete.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Autocomplete from '../../../react-component/container/input-wrappers/autocomplete'
+import Autocomplete from '../../../react-component/input-wrappers/autocomplete'
 
 const Root = styled.div`
   height: 100%;
