+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/checkbox/checkbox.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Checkbox from '../../../react-component/container/input-wrappers/checkbox'
+import Checkbox from '../../../react-component/input-wrappers/checkbox'
 
 const Root = styled.div`
   height: 100%;
