+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/color/color.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Color from '../../../react-component/container/input-wrappers/color'
+import Color from '../../../react-component/input-wrappers/color'
 
 const Root = styled.div`
   height: 100%;
