+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/multi-enum/multi-enum.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import MultiEnum from '../../../react-component/container/input-wrappers/multi-enum'
+import MultiEnum from '../../../react-component/input-wrappers/multi-enum'
 
 const Root = styled.div`
   height: 100%;
