+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/range/range.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Range from '../../../react-component/container/input-wrappers/range'
+import Range from '../../../react-component/input-wrappers/range'
 
 const Root = styled.div`
   height: 100%;
