+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/text/text.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Text from '../../../react-component/container/input-wrappers/text'
+import Text from '../../../react-component/input-wrappers/text'
 
 const Root = styled.div`
   height: 100%;
