+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/textarea/textarea.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Textarea from '../../../react-component/container/input-wrappers/textarea'
+import Textarea from '../../../react-component/input-wrappers/textarea'
 
 const Root = styled.div`
   height: 100%;
