+++ ui/packages/catalog-ui-search/src/main/webapp/dev/component/thumbnail/thumbnail.tsx
@@ -18,7 +18,7 @@ import { LiveProvider, LiveEditor, LiveError, LivePreview } from 'react-live'
 import { hot } from 'react-hot-loader'
 import alert from '../utils/alert'
 
-import Thumbnail from '../../../react-component/container/input-wrappers/thumbnail'
+import Thumbnail from '../../../react-component/input-wrappers/thumbnail'
 
 const Root = styled.div`
   height: 100%;
