+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/input-wrappers/base/base.tsx
@@ -12,14 +12,14 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-import { Subtract } from '../../../../typescript'
+import { Subtract } from '../../../typescript'
 
 import * as React from 'react'
-import MarionetteRegionContainer from '../../../marionette-region-container'
-import withListenTo, { WithBackboneProps } from '../../../backbone-container'
+import MarionetteRegionContainer from '../../marionette-region-container'
+import withListenTo, { WithBackboneProps } from '../../backbone-container'
 import { hot } from 'react-hot-loader'
-const PropertyView = require('../../../../component/property/property.view.js')
-const PropertyModel = require('../../../../component/property/property.js')
+const PropertyView = require('../../../component/property/property.view.js')
+const PropertyModel = require('../../../component/property/property.js')
 
 export enum Type {
   autocomplete = 'AUTOCOMPLETE',
