+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/input-wrappers/checkbox/index.tsx
