+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspaces-templates/presentation.tsx
@@ -17,7 +17,7 @@ import styled from 'styled-components'
 import { CustomElement } from '../styles/mixins'
 import { ChangeBackground } from '../styles/mixins'
 import { transparentize, readableColor } from 'polished'
-import Text from '../container/input-wrappers/text'
+import Text from '../input-wrappers/text'
 
 type Props = {
   startAdhocSearch: () => void
